+++ src/modules/legend/tests/unit/components/LegendWindow.spec.js
@@ -263,5 +263,77 @@ describe("LegendWindow.vue", () => {
                 expect(wrapper.vm.prepareLegendForGroupLayer(layerSource)).to.deep.equal(["foobar", "barfoo", "barbar", "foofoo"]);
             });
         });
+        describe("prepareLegendForPoint", () => {
+            it("prepareLegendForPoint - imageName contains whole path", () => {
+                const legendObj = {},
+                    style = {
+                        imagePath: "foo/bar/",
+                        type: "icon",
+                        imageName: "https://localhost:9001/portal/basic/geodaten/icons/svg/poi/Rathaus.svg"
+                    };
+                let result = null;
+
+                wrapper = shallowMount(LegendWindowComponent, {store, localVue});
+                result = wrapper.vm.prepareLegendForPoint(legendObj, style);
+
+                expect(result.graphic).to.deep.equal(style.imageName);
+            });
+            it("prepareLegendForPoint - imageName and imagePath contain whole path", () => {
+                const legendObj = {},
+                    style = {
+                        imagePath: "https://localhost:9001/portal/basic/geodaten2/icons",
+                        type: "icon",
+                        imageName: "https://localhost:9001/portal/basic/geodaten/icons/svg/poi/Rathaus.svg"
+                    };
+                let result = null;
+
+                wrapper = shallowMount(LegendWindowComponent, {store, localVue});
+                result = wrapper.vm.prepareLegendForPoint(legendObj, style);
+
+                expect(result.graphic).to.deep.equal(style.imageName);
+            });
+            it("prepareLegendForPoint - imageName contains complete svg", () => {
+                const legendObj = {},
+                    style = {
+                        imagePath: "foo/bar/",
+                        type: "icon",
+                        imageName: "<svg xmlns='http://www.w3.org/2000/svg' width='16' height='16' fill='#E10019' class='bi bi-geo-fill' viewBox='0 0 16 16'><path fill-rule='evenodd' d='M4 4a4 4 0 1 1 4.5 3.969V13.5a.5.5 0 0 1-1 0V7.97A4 4 0 0 1 4 3.999zm2.493 8.574a.5.5 0 0 1-.411.575c-.712.118-1.28.295-1.655.493a1.319 1.319 0 0 0-.37.265.301.301 0 0 0-.057.09V14l.002.008a.147.147 0 0 0 .016.033.617.617 0 0 0 .145.15c.165.13.435.27.813.395.751.25 1.82.414 3.024.414s2.273-.163 3.024-.414c.378-.126.648-.265.813-.395a.619.619 0 0 0 .146-.15.148.148 0 0 0 .015-.033L12 14v-.004a.301.301 0 0 0-.057-.09 1.318 1.318 0 0 0-.37-.264c-.376-.198-.943-.375-1.655-.493a.5.5 0 1 1 .164-.986c.77.127 1.452.328 1.957.594C12.5 13 13 13.4 13 14c0 .426-.26.752-.544.977-.29.228-.68.413-1.116.558-.878.293-2.059.465-3.34.465-1.281 0-2.462-.172-3.34-.465-.436-.145-.826-.33-1.116-.558C3.26 14.752 3 14.426 3 14c0-.599.5-1 .961-1.243.505-.266 1.187-.467 1.957-.594a.5.5 0 0 1 .575.411z'/></svg>"
+                    };
+                let result = null;
+
+                wrapper = shallowMount(LegendWindowComponent, {store, localVue});
+                result = wrapper.vm.prepareLegendForPoint(legendObj, style);
+
+                expect(result.graphic).to.deep.equal("data:image/svg+xml;charset=utf-8," + encodeURIComponent(style.imageName));
+            });
+            it("prepareLegendForPoint - imageName contains path", () => {
+                const legendObj = {},
+                    style = {
+                        imagePath: "foo/bar/",
+                        type: "icon",
+                        imageName: "/svg/poi/Rathaus.svg"
+                    };
+                let result = null;
+
+                wrapper = shallowMount(LegendWindowComponent, {store, localVue});
+                result = wrapper.vm.prepareLegendForPoint(legendObj, style);
+
+                expect(result.graphic).to.deep.equal(style.imagePath + style.imageName);
+            });
+            it("prepareLegendForPoint - imageName contains icon name", () => {
+                const legendObj = {},
+                    style = {
+                        imagePath: "./foo/bar/",
+                        type: "icon",
+                        imageName: "Rathaus.svg"
+                    };
+                let result = null;
+
+                wrapper = shallowMount(LegendWindowComponent, {store, localVue});
+                result = wrapper.vm.prepareLegendForPoint(legendObj, style);
+
+                expect(result.graphic).to.deep.equal(style.imagePath + style.imageName);
+            });
+        });
     });
 });