+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 
 ### Fixed
 - addWMS: Transformation for external services with EPSG:4326
+- Issue #1052 VectorStyle: path to an icon in the map or in the legend in style.json may be an icon name, a relative path to icon, a complete url or an svg.
 - Issue #1018:The tool selectFeatures can now also be used mobile, as an area can be dragged on the map.
 - Issue #1083: fixed the issue.
 - Issue #1086: prepush hook runs on linux.